fix gcc-3.3 warnings on Solaris 8
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@984 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ server/gfmd/gfmd.c
@@ -803,7 +803,7 @@ main(int argc, char **argv)
 		 * We do this after calling gfarm_daemon(),
 		 * because it changes pid.
 		 */
-		fprintf(pid_fp, "%d\n", getpid());
+		fprintf(pid_fp, "%ld\n", (long)getpid());
 		fclose(pid_fp);
 	}
